Fold the interpreter class, registry and shared helpers into interpreter.ts
Only the per-command handlers keep their own files under
interpreter/commands/. To avoid a circular runtime dependency
(interpreter.ts imports the handlers to build its registry), breakPath
and resolvePosition are now methods on InterpreterContext, so handlers
only need a type-only import of CommandHandler that is erased at
runtime.
